another

+++ Task3/src/main/java/com/ties456/service/director/DirectorService.java
@@ -6,21 +6,21 @@ import com.ties456.model.director.Director;
 import java.util.List;
 
 /**
- * Created by V.Tsybulko on 30.09.2016.
+ * Created by chinhnk on 10/2/2016.
  */
 public interface DirectorService {
 
     List<Director> getAll();
 
-    Director getById(long id);
+    Director getById(int id);
 
-    boolean isDirectorExist(long id);
+    boolean isDirectorExist(int id);
 
     Director saveDirector(Director person);
 
     void updateDirector(Director person);
 
-    void deleteDirectorById(long id);
+    void deleteDirectorById(int id);
 
     void deleteAllDirectors();
