+++ Task3/src/main/java/com/ties456/service/movie/MovieService.java
@@ -5,21 +5,21 @@ import com.ties456.model.movie.Movie;
 import java.util.List;
 
 /**
- * Created by V.Tsybulko on 30.09.2016.
+ * Created by chinhnk on 10/2/2016.
  */
 public interface MovieService {
 
     List<Movie> getAll();
 
-    Movie getById(long id);
+    Movie getById(int id);
 
-    boolean isMovieExist(long id);
+    boolean isMovieExist(int id);
 
     Movie saveMovie(Movie movie);
 
     void updateMovie(Movie movie);
 
-    void deleteMovieById(long id);
+    void deleteMovieById(int id);
 
     void deleteAllMovies();
